+++ app/worker/worker.go
@@ -1,12 +1,13 @@
 package worker
 
 import (
+	"context"
 	"math/big"
-	
+
 	"github.com/spf13/viper"
 	commonHandlers "gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/common/handlers"
-	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/core/worker"
 	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/core/services"
+	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/core/worker"
 	"gitlab.com/ConsenSys/client/fr/core-stack/worker/tx-crafter.git/app/infra"
 	"gitlab.com/ConsenSys/client/fr/core-stack/worker/tx-crafter.git/app/worker/handlers"
 )
@@ -14,7 +15,7 @@ import (
 // CreateWorker creates worker and attach it to application
 func CreateWorker(infra *infra.Infra, marker services.OffsetMarker) *worker.Worker {
 	// Instantiate worker
-	w := worker.NewWorker(worker.NewConfig())
+	w := worker.NewWorker(context.Background(), worker.NewConfig())
 
 	// Handler::loader
 	w.Use(commonHandlers.Loader(infra.Unmarshaller))